FoolOfRESTAPI/TelegramLogger/pyproject.toml
[project]
name = "telegramlogger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.9.0",
    "certifi==2025.6.15",
    "dotenv==0.9.9",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "psycopg==3.2.9",
    "psycopg-binary>=3.3.3",
    "python-dotenv==1.1.1",
    "python-telegram-bot==22.2",
    "sniffio==1.3.1",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
]
